requirements_chromadb_uploader.txt
streamlit>=1.28.0
pandas>=1.5.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
tqdm>=4.64.0
python-dotenv>=1.0.0
